Update config.lua
at the moment, it stores all money files directly in the world folder. this makes it very hard to work with other files especially once your player list starts to get bigger. so this change is part of what will allow us to define where we store the money files.